Extract menu page to separete directory
