+++ lib/pages/menu/menu_item_list_view.dart
@@ -7,7 +7,7 @@ import 'package:url_launcher/url_launcher.dart';
 import 'package:pola_flutter/analytics/analytics_about_row.dart';
 import 'package:pola_flutter/analytics/pola_analytics.dart';
 import 'package:pola_flutter/i18n/strings.g.dart';
-import 'web_view_dialog.dart';
+import '../../ui/web_view_dialog.dart';
 
 class MenuItemListView extends StatelessWidget {
   final PolaAnalytics analytics;
